+++ llvm/tools/llvm-symbolizer/llvm-symbolizer.cpp
@@ -32,6 +32,7 @@
 #include "llvm/Support/COM.h"
 #include "llvm/Support/CommandLine.h"
 #include "llvm/Support/Debug.h"
+#include "llvm/Support/Errc.h"
 #include "llvm/Support/FileSystem.h"
 #include "llvm/Support/InitLLVM.h"
 #include "llvm/Support/Path.h"
@@ -86,10 +87,10 @@ public:
 
 static std::string ToolName;
 
-static void printError(const ErrorInfoBase &EI, StringRef Path) {
+static void printError(const ErrorInfoBase &EI, StringRef AuxInfo) {
   WithColor::error(errs(), ToolName);
-  if (!EI.isA<FileError>())
-    errs() << "'" << Path << "': ";
+  if (!AuxInfo.empty())
+    errs() << "'" << AuxInfo << "': ";
   EI.log(errs());
   errs() << '\n';
 }
@@ -157,10 +158,14 @@ static StringRef getSpaceDelimitedWord(StringRef &Source) {
   return Result;
 }
 
-static bool parseCommand(StringRef BinaryName, bool IsAddr2Line,
-                         StringRef InputString, Command &Cmd,
-                         std::string &ModuleName, object::BuildID &BuildID,
-                         uint64_t &ModuleOffset) {
+static Error getStringError(StringRef Msg) {
+  return make_error<StringError>(Msg, inconvertibleErrorCode());
+}
+
+static Error parseCommand(StringRef BinaryName, bool IsAddr2Line,
+                          StringRef InputString, Command &Cmd,
+                          std::string &ModuleName, object::BuildID &BuildID,
+                          uint64_t &ModuleOffset) {
   ModuleName = BinaryName;
   if (InputString.consume_front("CODE ")) {
     Cmd = Command::Code;
@@ -180,15 +185,13 @@ static bool parseCommand(StringRef BinaryName, bool IsAddr2Line,
     InputString = InputString.ltrim();
     if (InputString.consume_front("FILE:")) {
       if (HasFilePrefix || HasBuildIDPrefix)
-        // Input file specification prefix is already seen.
-        return false;
+        return getStringError("Duplicate input file specification prefix");
       HasFilePrefix = true;
       continue;
     }
     if (InputString.consume_front("BUILDID:")) {
       if (HasBuildIDPrefix || HasFilePrefix)
-        // Input file specification prefix is already seen.
-        return false;
+        return getStringError("Duplicate input file specification prefix");
       HasBuildIDPrefix = true;
       continue;
     }
@@ -196,17 +199,14 @@ static bool parseCommand(StringRef BinaryName, bool IsAddr2Line,
   }
   if (HasBuildIDPrefix || HasFilePrefix) {
     if (!BinaryName.empty() || !BuildID.empty())
-      // Input file is already specified in command line.
-      return false;
+      return getStringError("Input file is already specified");
     StringRef Name = getSpaceDelimitedWord(InputString);
     if (Name.empty())
-      // Wrong name for module file.
-      return false;
+      return getStringError("Input file name is incorrect");
     if (HasBuildIDPrefix) {
       BuildID = parseBuildID(Name);
       if (BuildID.empty())
-        // Wrong format of BuildID hash.
-        return false;
+        return getStringError("Wrong format of build-id");
     } else {
       ModuleName = Name;
     }
@@ -216,8 +216,7 @@ static bool parseCommand(StringRef BinaryName, bool IsAddr2Line,
       // items, assume that the first item is a file name.
       StringRef Name = getSpaceDelimitedWord(InputString);
       if (Name.empty() || InputString.empty())
-        // No input filename is specified.
-        return false;
+        return getStringError("No input filename is specified");
       ModuleName = Name;
     }
   }
@@ -230,7 +229,9 @@ static bool parseCommand(StringRef BinaryName, bool IsAddr2Line,
   // "0x" or "0X" prefix; do the same for compatibility.
   if (IsAddr2Line)
     Offset.consume_front("0x") || Offset.consume_front("0X");
-  return !Offset.getAsInteger(IsAddr2Line ? 16 : 0, ModuleOffset);
+  if (Offset.getAsInteger(IsAddr2Line ? 16 : 0, ModuleOffset))
+    return errorCodeToError(errc::invalid_argument);
+  return Error::success();
 }
 
 template <typename T>
@@ -285,9 +286,16 @@ static void symbolizeInput(const opt::InputArgList &Args,
   std::string ModuleName;
   object::BuildID BuildID(IncomingBuildID.begin(), IncomingBuildID.end());
   uint64_t Offset = 0;
-  if (!parseCommand(Args.getLastArgValue(OPT_obj_EQ), IsAddr2Line,
-                    StringRef(InputString), Cmd, ModuleName, BuildID, Offset)) {
-    Printer.printInvalidCommand({ModuleName, std::nullopt}, InputString);
+  if (Error E = parseCommand(Args.getLastArgValue(OPT_obj_EQ), IsAddr2Line,
+                             StringRef(InputString), Cmd, ModuleName, BuildID,
+                             Offset)) {
+    handleAllErrors(
+        std::move(E),
+        [&](const StringError &EI) { printError(EI, InputString); },
+        [&](const ECError &EI) {
+          Request SymRequest = {ModuleName};
+          Printer.print(SymRequest, DILineInfo());
+        });
     return;
   }
   bool ShouldInline = Args.hasFlag(OPT_inlines, OPT_no_inlines, !IsAddr2Line);