+++ llvm/docs/ReleaseNotes.rst
@@ -132,6 +132,10 @@ Changes to the Debug Info
 Changes to the LLVM tools
 ---------------------------------
 
+* llvm-symbolizer now treats invalid input as an address for which source
+  information is not found.
+
+
 Changes to LLDB
 ---------------------------------